==> notebooks/01_data_pull.ipynb <==
!pip install "osmnx>=1.8.0,<2.0.0"


import osmnx as ox
import random
import pandas as pd

# More reliable boundary for routing
place = "Atlanta, Georgia, USA"
G = ox.graph_from_place(place, network_type="drive")

nodes = list(G.nodes)
stops = random.sample(nodes, 100)

locations = [(G.nodes[n]['y'], G.nodes[n]['x']) for n in stops]
df_stops = pd.DataFrame(locations, columns=["latitude", "longitude"])
df_stops.to_csv("../data/atlanta_stops.csv", index=False)

df_stops.head(), len(df_stops)
